--- lib/utilities/etl.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require "utilities/etl/raw_member_row"
-require "utilities/etl/extract_member_data"
-require "utilities/etl/import_manager"
