--- lib/utilities/etl/raw_member_row.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module Etl
-  RawMemberRow = Struct.new(:name, :gender, :birthdate, :phone_number, :email, keyword_init: true)
-end
